Add a /file/<id>/extend API to extend an existing upload expiry
Requires no arguments, but you can provide a X-FS-TTL header that works
the same way it does on the upload endpoint (i.e. to override the
default upload period).

The endpoint will not reduce an existing expiry but rather leaves it
alone.  If the file was found it returns the same file info as the
/file/<id>/info endpoint (including the new expiry timestamp); returns
404 if not found.