Use Ed base key for fileserver, with X keys for onion requests
Onion requests use X keys for shared secret generation, which causes
problems with future functionality (lokinet, quic) that expects Ed25519
pubkeys.

This changes file server to use an Ed25519 keypair under the hood so
that it can still do onion requests (using the Ed -> X converted keys),
but that we can switch to using the master Ed25519 key everywhere and
switch to X only when needed.